Add css var scroll
